fix Horizontal Privilege Escalation and prompt password for backup
